+++ backend/hexa/pipelines/models.py
@@ -5,12 +5,13 @@ import typing
 import uuid
 from datetime import datetime
 
+from croniter import croniter
 from django.apps import apps
 from django.conf import settings
 from django.contrib.auth.models import AnonymousUser
 from django.contrib.contenttypes.fields import GenericRelation
 from django.contrib.postgres.indexes import GinIndex, GistIndex
-from django.core.exceptions import PermissionDenied
+from django.core.exceptions import PermissionDenied, ValidationError
 from django.core.signing import Signer, TimestampSigner
 from django.db import models, transaction
 from django.db.models import Q
@@ -479,22 +480,16 @@ class Pipeline(SoftDeletedModel):
         """
         Get the config from the previous version of the pipeline considering only overlapping parameters between the new and the previous version.
         """
-
-        def remove_default(parameter):
-            parameter_without_default = parameter.copy()
-            if "default" in parameter:
-                del parameter_without_default["default"]
-            return parameter_without_default
-
         previous_config_from_overlapping_parameters = {}
         if self.last_version:
-            previous_parameters = list(
-                map(remove_default, self.last_version.parameters)
-            )
+            previous_param_keys = {
+                (p["code"], p.get("type")) for p in self.last_version.parameters
+            }
             overlapping_parameters = [
                 new_parameter
                 for new_parameter in new_parameters
-                if remove_default(new_parameter) in previous_parameters
+                if (new_parameter["code"], new_parameter.get("type"))
+                in previous_param_keys
             ]
             previous_param_default_by_code = {
                 previous_parameter["code"]: previous_parameter.get("default")
@@ -571,9 +566,12 @@ class Pipeline(SoftDeletedModel):
         ):
             raise MissingPipelineConfiguration
 
-        for key in ["name", "description", "schedule", "config", "functional_type"]:
+        for key in ["name", "description", "config", "functional_type"]:
             if key in kwargs:
                 setattr(self, key, kwargs[key])
+
+        if "schedule" in kwargs:
+            self.set_schedule(kwargs["schedule"])
 
         if "auto_update_from_template" in kwargs:
             self.auto_update_from_template = kwargs["auto_update_from_template"]
@@ -621,6 +619,12 @@ class Pipeline(SoftDeletedModel):
             self.generate_webhook_token()
 
         self.webhook_enabled = enabled
+
+    def set_schedule(self, schedule: str):
+        if not croniter.is_valid(schedule):
+            raise ValidationError("Invalid cron expression")
+        else:
+            self.schedule = schedule
 
     def generate_webhook_token(self):
         signer = TimestampSigner()